Restructure Search components to allow for more flexibility
