--- src/components/Search/index.js
@@ -1,5 +1,0 @@
-import SearchPresentation from 'components/Search/SearchPresentation'
-
-export default {
-  Presentation: SearchPresentation,
-}
